Update skeleton-application.md
I've tried on a few occasions to follow this outdated tutorial. It is a constant battle trying to get advanced developers to support documentation.  Please hire someone to do this or prioritize this task otherwise people will just use Laravel which has a lot more support and clearer directions. 

Additionally, Depending on the answer to the prompt you don't mention there is an error code 1 that happens. You might want to resolve that too. Finally, your sqlite instructions don't work. If my job wasn't using this framework I'd disregard it altogether...this is so frustrating.